some casts \?

+++ trunk/src/kernel/gmp++/gmp++_int_cstor.C
@@ -69,7 +69,7 @@ namespace Givaro {
 	// log[10](2^8) < 2.408239966
 	Integer::Integer(long long int n)
 	{
-		char * tmp = new char[long(2.408239966*sizeof(long long))+1]; sprintf(tmp,"%lld",n);
+		char * tmp = new char[long(2.408239966*(double)sizeof(long long))+1]; sprintf(tmp,"%lld",n);
 		mpz_init_set_str((mpz_ptr)&gmp_rep, tmp, 10) ;
 		delete [] tmp;
 	}
